feat: move SearchServicesViewModel to its own file
