+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/search/SearchServicesViewModel.kt
@@ -1,0 +1,42 @@
+/*
+ * Wire
+ * Copyright (C) 2023 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ */
+package com.wire.android.ui.home.conversations.search
+
+import com.wire.android.navigation.NavigationManager
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.onStart
+
+@OptIn(ExperimentalCoroutinesApi::class)
+abstract class SearchServicesViewModel(
+    navigationManager: NavigationManager
+) : SearchPeopleViewModel(
+    navigationManager = navigationManager
+) {
+
+    protected val servicesSearchQueryFlow = mutableSearchQueryFlow
+        .flatMapLatest { searchTerm ->
+            searchServices(searchTerm)
+                .onStart {
+                    emit(SearchResultState.InProgress)
+                }
+        }
+
+    abstract suspend fun searchServices(searchTerm: String): Flow<SearchResultState>
+}
